Refactor user creation and password change DTOs; update related imports and validation logic
