+++ src/app/dashboard/hooks/useUsuario.ts
@@ -2,11 +2,9 @@ import { AxiosError } from "axios";
 import useSWRMutation from "swr/mutation";
 import { httpBase } from "../../config/api-base";
 import useSWR from "swr";
-import { ListarProveedoresDTO } from "@/app/api/proveedor/application/dto/listarProveedore.DTO";
-import { CrearProveedorDTO } from "@/app/api/proveedor/application/dto/crearProveedorDto";
-import { EditarProveedorDTO } from "../../api/proveedor/application/dto/editarProveedorDTO";
 import { ListarUsuriosDTO } from "@/app/api/usuarios/use-cases/dto/listarUsuarios.DTO";
-import { CambiarPasswordDTO } from "@/app/api/usuarios/use-cases/dto/crearUsuario.DTO";
+import { CambiarPasswordDTO } from "@/app/api/usuarios/use-cases/dto/cambiarPasswordDTO";
+import { CrearUsuarioDTO } from "@/app/api/usuarios/use-cases/dto/crearUsuario.DTO";
 
 export const useListadoUsuarios = () => {
   const fetcher = (url: string) => httpBase.get(url).then((res) => res.data);
@@ -43,7 +41,7 @@ export const useObtenerUsuarios = () => {
   };
 };
 export const useCrearUsuario = () => {
-  const fetcher = (url: string, { arg }: { arg: CambiarPasswordDTO }) =>
+  const fetcher = (url: string, { arg }: { arg: CrearUsuarioDTO }) =>
     httpBase.post(url, arg).then((res) => res.data);
 
   const { data, error, trigger, isMutating } = useSWRMutation(
